+++ packages/schema-to-zod/CHANGELOG.md
@@ -1,5 +1,12 @@
 # orchid-orm-schema-to-zod
 
+## 0.3.19
+
+### Patch Changes
+
+- Updated dependencies
+  - pqb@0.10.18
+
 ## 0.3.18
 
 ### Patch Changes